Switch to using appropriate libvolk functions instead of trivial
for-loops. This should improve performance and reduce cpu load in some cases.

Requires https://github.com/gqrx-sdr/gqrx/pull/1042