@argilo 2022-01-11
:thinking: So far I have not been able to observe a performance improvement on this branch. In fact, it looks like master might perform a bit better. I suppose it could be that the compiler is already applying SIMD optimizations to the loops.