@argilo 2022-01-11
Have you been able to measure any performance improvement on your system? We probably shouldn't merge this until it's clear that there's a benefit.