@vladisslav2011 2022-01-11
> Have you been able to measure any performance improvement on your system?

I see small improvement on Intel Atom N270 system.
128k FFT, 10 fps update rate, maximum zoom, no demodulator
master ~54.5% CPU consumed by main thread
this branch ~53.7% CPU consumed by main thread
Less than 1% difference. I think GCC autovectorization works very well on x86.
It may be better to test on ARM, but I do not own a GUI-capable ARM system.
There shouldn't be much improvement as application of the window takes much less time, than actual fft calculation and even less time, than rendering waterfall/spectrum.