@argilo 2022-01-12
Since the performance difference was so small, I made more precise measurements by placing `std::chrono::high_resolution_clock` around just the lines that were changed, and averaged over many executions.

On my laptop, which has a i7-8550U CPU, I got the following results:

rx_meter_c, used with wideband FM (240 ksps):
* before: 0.232 ms
* after: 0.048 ms

This gives a savings of 0.184 ms per execution. At 10 executions per second, the total reduction in CPU utilization should be about 0.184%.

rx_fft_c, used with 1048576-point FFT:
* before: 1.965 ms
* after: 1.972 ms

Presumably the compiler already optimized this simple loop to use SIMD instructions.

These results would explain why the performance difference was not visible using top.