@argilo 2022-01-12
Since there does appear to be a (very small) performance improvement, I'll merge this once GitHub Actions is working again and CI passes.

In the future, it would be very helpful if you would benchmark potential performance improvements so that it is easier to decide whether they are worthwhile.