@argilo 2022-01-12
For reference, I used more-or-less the code from this answer: https://stackoverflow.com/questions/22387586/measuring-execution-time-of-a-function-in-c